Aula44 - Exclusao da foto do usuario

+++ application/models/usuarios_model.php
@@ -36,6 +36,13 @@ class Usuarios_model extends CI_Model {
     }
 
     function deletar($idusuario) {
+       $busca = $this->editar($idusuario);
+       $caminho = "c:/xampp/htdocs/sitApTecVesp2014/assets/images/" ;
+       $arquivo = $caminho . $busca[0]->foto;
+       
+       if(is_file($arquivo)) unlink ($arquivo);
+       
+       
         $this->db->where('idusuario', $idusuario);
         return $this->db->delete('usuario');
     }
